xe: refactor gemmstone includes to be independent of oneDNN structure
