+++ src/gpu/intel/jit/gemm/gen_gemm_kernel.cpp
@@ -16,10 +16,10 @@
 
 #include "gpu/intel/jit/gemm/gen_gemm_kernel.hpp"
 #include "common/impl_registration.hpp"
+#include "gemmstone/generator.hpp"
+#include "gemmstone/strategy_parser.hpp"
 #include "gpu/intel/compute/device_info.hpp"
 #include "gpu/intel/jit/gemm/gen_gemm_kernel_db.hpp"
-#include "gpu/intel/jit/gemm/include/generator.hpp"
-#include "gpu/intel/jit/gemm/include/strategy_parser.hpp"
 #include "gpu/intel/jit/utils/ngen_type_bridge.hpp"
 #include "gpu/intel/utils.hpp"
