+++ src/gpu/intel/jit/gemm/gen_gemm_kernel.hpp
@@ -18,15 +18,15 @@
 #define GPU_INTEL_JIT_GEMM_GEN_GEMM_KERNEL_HPP
 
 #include "common/c_types_map.hpp"
+#include "gemmstone/driver_info.hpp"
+#include "gemmstone/kernel_catalog.hpp"
+#include "gemmstone/kernel_evaluator.hpp"
+#include "gemmstone/kernel_selector.hpp"
+#include "gemmstone/problem.hpp"
+#include "gemmstone/strategy.hpp"
+#include "gemmstone/type.hpp"
 #include "gpu/intel/compute/device_info.hpp"
 #include "gpu/intel/compute/kernel_arg_list.hpp"
-#include "gpu/intel/jit/gemm/include/driver_info.hpp"
-#include "gpu/intel/jit/gemm/include/kernel_catalog.hpp"
-#include "gpu/intel/jit/gemm/include/kernel_evaluator.hpp"
-#include "gpu/intel/jit/gemm/include/kernel_selector.hpp"
-#include "gpu/intel/jit/gemm/include/problem.hpp"
-#include "gpu/intel/jit/gemm/include/strategy.hpp"
-#include "gpu/intel/jit/gemm/include/type.hpp"
 #include "gpu/intel/jit/generator_base.hpp"
 #include "gpu/intel/kernel_cache.hpp"
 #include "xpu/utils.hpp"
