+++ src/gpu/intel/jit/gemm/include/internal/ngen_includes.hpp
@@ -14,7 +14,7 @@
 * limitations under the License.
 *******************************************************************************/
 
-#include "../config.hpp"
+#include "gemmstone/config.hpp"
 
 #   include "ngen_opencl.hpp"
 #   include "ngen_register_allocator.hpp"
