+++ src/gpu/intel/jit/gemm/xe_hp_systolic_gemm.hpp
@@ -24,9 +24,9 @@
 #include "common/c_types_map.hpp"
 #include "common/memory_storage.hpp"
 #include "common/utils.hpp"
+#include "gemmstone/driver_info.hpp"
+#include "gemmstone/problem.hpp"
 #include "gpu/intel/gemm/gpu_gemm.hpp"
-#include "gpu/intel/jit/gemm/include/driver_info.hpp"
-#include "gpu/intel/jit/gemm/include/problem.hpp"
 #include "gpu/intel/jit/gemm/jit_gemm_pd.hpp"
 
 namespace dnnl {
