+++ src/gpu/intel/ocl/micro_sdpa.cpp
@@ -20,9 +20,9 @@
 #include "common/c_types_map.hpp"
 #include "common/sdpa_utils.hpp"
 #include "common/type_helpers.hpp"
+#include "gemmstone/microkernel_provider.hpp"
 #include "gpu/intel/compute/utils.hpp"
 #include "gpu/intel/jit/gemm/gen_gemm_kernel.hpp"
-#include "gpu/intel/jit/gemm/include/microkernel_provider.hpp"
 
 #include <algorithm>
 #include <cstdio>
